Fix issue where extension could delete all files
Throw if `storageUri` is not set instead of falling back to ''. `storageUri` is undefined if no folder is open in vscode (https://vscode-api.js.org/interfaces/vscode.ExtensionContext.html#storageUri).

Add an assertion in the deleteFile() utility method to ensure deletes only work on paths that include the name of the extension.

Fixes #48